fix: integration tests
